pyproject.toml
[project]
name = "document-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "google-genai>=1.66.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
]
